refactor: flows-test-kit mocks
